added email form field into login page
