--- client/src/components/Forms/Email/index.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-function Email() {
-  return (
-    <div>
-      <Form>
-        <Form.Group controlId="formBasicEmail">
-          <Form.Label>Email address</Form.Label>
-          <Form.Control type="email" placeholder="Enter email" />
-          <Form.Text className="text-muted">
-            We'll never share your email with anyone else.
-          </Form.Text>
-        </Form.Group>
-      </Form>
-    </div>
-  );
-}
-
-export default Email;
